refactor: 상수 제거, 비즈니스 로직과 UI 로직 분리
