+++ src/constants/index.ts
@@ -1,0 +1,7 @@
+const PASSENGER = {
+  DEFAULT: 1,
+  MIN: 0,
+  MAX: 3,
+};
+
+export default PASSENGER;
